Added RegionFactory to the default Ehcache hibernate integration. Plus tests. Upgraded Ehcache to 2.4.1
