+++ hibernate-ehcache/pom.xml
@@ -48,8 +48,8 @@
         </dependency>
         <dependency>
             <groupId>net.sf.ehcache</groupId>
-            <artifactId>ehcache</artifactId>
-            <version>1.5.0</version>
+            <artifactId>ehcache-core</artifactId>
+            <version>2.4.1</version>
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
