+++ hibernate-ehcache/src/test/java/org/hibernate/cache/EhCacheProviderTest.java
@@ -1,0 +1,64 @@
+/*
+ * Hibernate, Relational Persistence for Idiomatic Java
+ *
+ * Copyright (c) 2007, Red Hat Middleware LLC or third-party contributors as
+ * indicated by the @author tags or express copyright attribution
+ * statements applied by the authors.  All third-party contributions are
+ * distributed under license by Red Hat Middleware LLC.
+ *
+ * This copyrighted material is made available to anyone wishing to use, modify,
+ * copy, or redistribute it subject to the terms and conditions of the GNU
+ * Lesser General Public License, as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
+ * for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this distribution; if not, write to:
+ * Free Software Foundation, Inc.
+ * 51 Franklin Street, Fifth Floor
+ * Boston, MA  02110-1301  USA
+ */
+package org.hibernate.cache;
+
+import junit.framework.Test;
+
+import org.hibernate.cfg.Environment;
+import org.hibernate.testing.junit.functional.FunctionalTestClassTestSuite;
+import org.hibernate.testing.cache.BaseCacheProviderTestCase;
+
+/**
+ * @author Emmanuel Bernard
+ */
+public class EhCacheProviderTest extends BaseCacheProviderTestCase {
+	public EhCacheProviderTest(String x) {
+		super(x);
+	}
+
+	public static Test suite() {
+		return new FunctionalTestClassTestSuite(EhCacheProviderTest.class);
+	}
+
+	public String getCacheConcurrencyStrategy() {
+		return "read-write";
+	}
+
+	protected Class getCacheProvider() {
+		return EhCacheProvider.class;
+	}
+
+	protected String getConfigResourceKey() {
+		return Environment.CACHE_PROVIDER_CONFIG;
+	}
+
+	protected String getConfigResourceLocation() {
+		return "ehcache.xml";
+	}
+
+	protected boolean useTransactionManager() {
+		return false;
+	}
+
+}
